requirements.txt
ccxt==4.3.88
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.3.14b0
backtesting==0.3.3
fastapi>=0.111.0
uvicorn>=0.30.0
APScheduler>=3.10.4
SQLAlchemy>=2.0.29
aiosqlite>=0.20.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
pydantic>=2.7.0
httpx>=0.27.0
loguru>=0.7.2
